requirements.txt
#
####### requirements.txt #######
#
Pillow
cryptography
